+++ koku-frontend/src/app/user/private-appointment-details/private-appointment-details.component.ts
@@ -5,6 +5,11 @@ import {PreventLosingChangesService} from "../../prevent-losing-changes/prevent-
 import {PrivateAppointmentService} from "./private-appointment.service";
 import * as moment from "moment";
 import {SnackBarService} from "../../snackbar/snack-bar.service";
+import {
+  AlertDialogButtonConfig,
+  AlertDialogComponent,
+  AlertDialogData
+} from "../../alert-dialog/alert-dialog.component";
 
 export interface PrivateAppointmentDetailsData {
   privateAppointmentId?: number;
@@ -103,12 +108,36 @@ export class PrivateAppointmentDetailsComponent implements AfterViewInit {
   }
 
   delete(appointment: KokuDto.PrivateAppointmentDto) {
-    this.saving = true;
-    this.privateAppointmentService.deletePrivateAppointment(appointment).subscribe(() => {
-      this.dialogRef.close();
-      this.saving = false;
-    }, () => {
-      this.saving = false;
+    const dialogData: AlertDialogData = {
+      headline: 'Privattermin Löschen',
+      message: `Wollen Sie den Termin wirklich löschen?`,
+      buttons: [{
+        text: 'Abbrechen',
+        onClick: (mouseEvent: Event, button: AlertDialogButtonConfig, dialogRef: MatDialogRef<AlertDialogComponent>) => {
+          dialogRef.close();
+        }
+      }, {
+        text: 'Bestätigen',
+        onClick: (mouseEvent: Event, button: AlertDialogButtonConfig, dialogRef: MatDialogRef<AlertDialogComponent>) => {
+          button.loading = true;
+          this.privateAppointmentService.deletePrivateAppointment(appointment).subscribe(() => {
+            dialogRef.close();
+            this.dialogRef.close();
+          }, () => {
+            button.loading = false;
+          });
+        }
+      }]
+    };
+
+    this.dialog.open(AlertDialogComponent, {
+      data: dialogData,
+      width: '100%',
+      maxWidth: 700,
+      closeOnNavigation: false,
+      position: {
+        top: '20px'
+      }
     });
   }
